Extract the route-shape primitives into a Routing leaf module [#806][#790]
Kernel dissolution, penultimate step (Refs #790, Closes #806). RouteSuffix and
ChallengePath are the plugin's route-shape contract — the
{protocol}/{path-kind}/{provider} suffix reader and the new/legacy
challenge-path classifier. They are real code dependencies of registered
modules (Oidc/OidcCallbackPath calls RouteSuffix.TryRead;
Shared/ChallengeNewPathResolver calls ChallengePath.IsNewPath), so they cannot
move into the upcoming Http boundary module — module edges must never point at
the boundary. They become the Routing leaf:

- Api/Routing: RouteSuffix, ChallengePath. DAG: Routing is a leaf; Oidc gains
  the Routing edge (its callback-path reader).
- AuthPageCsp moves to Api/Shared — its sole consumer is Shared/FlowResponses,
  and a served-page helper is exactly Shared's charter.
- Tests mirror the moves: Tests/Routing/, AuthPageCspTests to Tests/Shared/.

A dedicated two-file leaf instead of folding into Net is deliberate: Net is
generic networking/SSRF infrastructure; the route shape is the plugin's own URL
contract — separate concerns, and the module map now documents exactly that.

The flat Api kernel is down to the pure web boundary (SSOController,
RequestHelpers, ProviderConnectionTester) — the Http module (#807) finishes the
dissolution.

No behavior change. Verified by the full build (--warnaserror, both TFMs, 0
warnings), the complete suite (1589 both TFMs) and the ABI-floor build.
