+++ SSO-Auth.Tests/Routing/ChallengePathTests.cs
@@ -1,3 +1,4 @@
+using Jellyfin.Plugin.SSO_Auth.Api.Routing;
 using Jellyfin.Plugin.SSO_Auth.Api;
 using Xunit;
