+++ SSO-Auth.Tests/Shared/AuthPageCspTests.cs
@@ -1,3 +1,4 @@
+using Jellyfin.Plugin.SSO_Auth.Api.Shared;
 using Jellyfin.Plugin.SSO_Auth.Api;
 using Xunit;
