+++ SSO-Auth/Api/Oidc/OidcCallbackPath.cs
@@ -1,4 +1,5 @@
 using System;
+using Jellyfin.Plugin.SSO_Auth.Api.Routing;
 
 #nullable enable
